+++ lib/src/logger.dart
@@ -1,3 +1,0 @@
-import 'package:logging/logging.dart';
-
-final logger = Logger('lottie');