+++ lib/src/lottie.dart
@@ -2,6 +2,7 @@ import 'dart:typed_data';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import '../lottie.dart';
+import 'composition.dart';
 import 'frame_rate.dart';
 import 'l.dart';
 import 'lottie_builder.dart';
@@ -50,12 +51,14 @@ class Lottie extends StatefulWidget {
     Key? key,
     AssetBundle? bundle,
     LottieFrameBuilder? frameBuilder,
+    ImageErrorWidgetBuilder? errorBuilder,
     double? width,
     double? height,
     BoxFit? fit,
     Alignment? alignment,
     String? package,
     bool? addRepaintBoundary,
+    WarningCallback? onWarning,
   }) =>
       LottieBuilder.asset(
         name,
@@ -71,12 +74,14 @@ class Lottie extends StatefulWidget {
         key: key,
         bundle: bundle,
         frameBuilder: frameBuilder,
+        errorBuilder: errorBuilder,
         width: width,
         height: height,
         fit: fit,
         alignment: alignment,
         package: package,
         addRepaintBoundary: addRepaintBoundary,
+        onWarning: onWarning,
       );
 
   /// Creates a widget that displays an [LottieComposition] obtained from a [File].
@@ -93,11 +98,13 @@ class Lottie extends StatefulWidget {
     void Function(LottieComposition)? onLoaded,
     Key? key,
     LottieFrameBuilder? frameBuilder,
+    ImageErrorWidgetBuilder? errorBuilder,
     double? width,
     double? height,
     BoxFit? fit,
     Alignment? alignment,
     bool? addRepaintBoundary,
+    WarningCallback? onWarning,
   }) =>
       LottieBuilder.file(
         file,
@@ -112,11 +119,13 @@ class Lottie extends StatefulWidget {
         onLoaded: onLoaded,
         key: key,
         frameBuilder: frameBuilder,
+        errorBuilder: errorBuilder,
         width: width,
         height: height,
         fit: fit,
         alignment: alignment,
         addRepaintBoundary: addRepaintBoundary,
+        onWarning: onWarning,
       );
 
   /// Creates a widget that displays an [LottieComposition] obtained from a [Uint8List].
@@ -133,11 +142,13 @@ class Lottie extends StatefulWidget {
     void Function(LottieComposition)? onLoaded,
     Key? key,
     LottieFrameBuilder? frameBuilder,
+    ImageErrorWidgetBuilder? errorBuilder,
     double? width,
     double? height,
     BoxFit? fit,
     Alignment? alignment,
     bool? addRepaintBoundary,
+    WarningCallback? onWarning,
   }) =>
       LottieBuilder.memory(
         bytes,
@@ -152,11 +163,13 @@ class Lottie extends StatefulWidget {
         onLoaded: onLoaded,
         key: key,
         frameBuilder: frameBuilder,
+        errorBuilder: errorBuilder,
         width: width,
         height: height,
         fit: fit,
         alignment: alignment,
         addRepaintBoundary: addRepaintBoundary,
+        onWarning: onWarning,
       );
 
   /// Creates a widget that displays an [LottieComposition] obtained from the network.
@@ -173,11 +186,13 @@ class Lottie extends StatefulWidget {
     void Function(LottieComposition)? onLoaded,
     Key? key,
     LottieFrameBuilder? frameBuilder,
+    ImageErrorWidgetBuilder? errorBuilder,
     double? width,
     double? height,
     BoxFit? fit,
     Alignment? alignment,
     bool? addRepaintBoundary,
+    WarningCallback? onWarning,
   }) =>
       LottieBuilder.network(
         url,
@@ -192,11 +207,13 @@ class Lottie extends StatefulWidget {
         onLoaded: onLoaded,
         key: key,
         frameBuilder: frameBuilder,
+        errorBuilder: errorBuilder,
         width: width,
         height: height,
         fit: fit,
         alignment: alignment,
         addRepaintBoundary: addRepaintBoundary,
+        onWarning: onWarning,
       );
 
   /// The Lottie composition to animate.