+++ lib/src/model/content/merge_paths.dart
@@ -1,6 +1,5 @@
 import '../../animation/content/content.dart';
 import '../../animation/content/merge_paths_content.dart';
-import '../../logger.dart';
 import '../../lottie_drawable.dart';
 import '../layer/base_layer.dart';
 import 'content_model.dart';
@@ -17,7 +16,8 @@ class MergePaths implements ContentModel {
   @override
   Content? toContent(LottieDrawable drawable, BaseLayer layer) {
     if (!drawable.enableMergePaths) {
-      logger.warning('Animation contains merge paths but they are disabled.');
+      drawable.composition
+          .addWarning('Animation contains merge paths but they are disabled.');
       return null;
     }
     return MergePathsContent(this);